Make it easier to subclass TransportCallTO + service
Move most of the complex logic from TransportCallTOService into an
abstract subclass.  This will enable UI-Support to make its own
extension of TransportCallTO + related services where it can get more
data along with the transport call.
