+++ src/main/java/org/dcsa/core/events/repository/AbstractTransportCallTORepository.java
@@ -3,6 +3,5 @@ package org.dcsa.core.events.repository;
 import org.dcsa.core.events.model.transferobjects.TransportCallTO;
 import org.dcsa.core.repository.ExtendedRepository;
 
-public interface TransportCallTORepository extends AbstractTransportCallTORepository<TransportCallTO> {
-
+public interface AbstractTransportCallTORepository<T extends TransportCallTO> extends ExtendedRepository<T, String> {
 }
