+++ src/main/java/org/dcsa/core/events/service/AbstractTransportCallTOService.java
@@ -1,0 +1,8 @@
+package org.dcsa.core.events.service;
+
+import org.dcsa.core.events.model.transferobjects.TransportCallTO;
+import org.dcsa.core.service.ExtendedBaseService;
+
+public interface AbstractTransportCallTOService<T extends TransportCallTO> extends ExtendedBaseService<T, String> {
+
+}
