+++ src/main/java/org/dcsa/core/events/service/TransportCallTOService.java
@@ -1,9 +1,6 @@
 package org.dcsa.core.events.service;
 
 import org.dcsa.core.events.model.transferobjects.TransportCallTO;
-import org.dcsa.core.service.BaseService;
-import org.dcsa.core.service.ExtendedBaseService;
 
-public interface TransportCallTOService extends ExtendedBaseService<TransportCallTO, String> {
-
+public interface TransportCallTOService extends AbstractTransportCallTOService<TransportCallTO> {
 }
